IO: split get_fields function to subclass specific ones
